Rename and restructure posts for navigation and readability
Renamed pages 'polizei.md' and 'brot.md' to align with posts' naming conventions. This helps to better organize the content and enhance search-ability across the website. The permalink of 'polizei.md' was also changed for relevance.

Additionally, improved readability of 'brot.md' by restructuring the content. This makeover introduces a more reader-friendly pattern that separates ingredients from instructions, making it easier for users to follow the bread recipe.