Update software development tools and libraries list
This commit updates the list of software development tools and libraries to reflect the use of the most current versions. This included adding the current Ubuntu OS, updating the version of PhpStorm and DDEV, as well as streamlining the list of PHP libraries. Redundant or outdated references were removed for a cleaner, more future-proofed list.